Refactor ConfiguratorHandlers.ts
Split `ConfiguratorHandlers.ts` into smaller, more manageable files while preserving functionality.
